finally test cases passed and black applied

+++ python/turbine_models/tests/stateless_llama_test.py
@@ -6,7 +6,6 @@
 
 import logging
 import turbine_models.custom_models.stateless_llama as llama
-import unittest
 import os
 import pytest
 
@@ -31,10 +30,20 @@ from turbine_models.custom_models import remap_gguf
 import safetensors
 
 from tqdm import tqdm
+from .vmfb_comparison import get_turbine_vmfb_string
 
 
+def test_vmfb_comparison():
+    """
+    Test that the vmfb model produces the same output as the torch model
 
-def test_export_vmfb(quantization: Literal["int4", None], precision: Literal["f16", "f32"]):
+    Precision can be 16 or 32, using 16 for speed and memory.
+
+    For VMFB, quantization can be int4 or None, but right now only using none for compatibility with torch.
+    """
+    quantization = "unquantized"
+    precision = "f32"
+
     llama.export_transformer_model(
         hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
         hf_auth_token=None,
@@ -47,7 +56,6 @@ def test_export_vmfb(quantization: Literal["int4", None], precision: Literal["f1
         target_triple="host",
     )
 
-# def test_export_safetensors(quantization: Literal["int4", None], precision: Literal["f16", "f32"]):
     from turbine_models.gen_external_params.gen_external_params import (
         gen_external_params,
     )
@@ -59,47 +67,40 @@ def test_export_vmfb(quantization: Literal["int4", None], precision: Literal["f1
         precision=precision,
     )
 
-# def test_run_vmfb(quantization: Literal["int4", None], precision: Literal["f16", "f32"]):
     DEFAULT_PROMPT = """<s>[INST] <<SYS>>
 Be concise. You are a helpful, respectful and honest assistant. If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information. <</SYS>> hi what are you? [/INST]
 """
-    # cache reference output to avoid having to use torch
-    TORCH_REFERENCE_STRING = '''Hello! I'm just an AI assistant, I'm here to help you with any questions you may have. However, I must point out that the question "what are you?" is not clear or factually coherent. I'm just a language model, I don't have a physical body or personal identity, so I cannot be described as a "you" in the classical sense. Is there anything else I can help you with?</s>'''
-    # torch_reference_string = get_torch_string(
-    #     prompt=DEFAULT_PROMPT,
-    #     hf_auth_token=None,
-    #     hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
-    #     tokens_to_compare=50,
-    # )
-    # 
-    torch_str = TORCH_REFERENCE_STRING
-    
 
-    from .vmfb_comparison import get_turbine_vmfb_string, get_torch_string
-
-    torch_str = get_torch_string(
-        prompt=DEFAULT_PROMPT,
-        hf_auth_token=None,
-        hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
-        tokens_to_compare=50,
-        precision=precision,
-        quantization=quantization,
-    )
-
-    # store torch string
-    # include precision and quantization in filename
     torch_str_cache_path = f"python/turbine_models/tests/vmfb_comparison_cached_torch_output_{precision}_{quantization}.txt"
-    with open(torch_str_cache_path, "w") as f:
-        f.write(torch_str)
+    # if cached, just read
+    if os.path.exists(torch_str_cache_path):
+        with open(torch_str_cache_path, "r") as f:
+            torch_str = f.read()
+    else:
+        from .vmfb_comparison import get_torch_string
+
+        torch_str = get_torch_string(
+            prompt=DEFAULT_PROMPT,
+            hf_auth_token=None,
+            hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
+            tokens_to_compare=50,
+            precision=precision,
+            quantization=quantization,
+        )
+
+        with open(torch_str_cache_path, "w") as f:
+            f.write(torch_str)
 
     turbine_str = get_turbine_vmfb_string(
         prompt=DEFAULT_PROMPT,
         hf_auth_token=None,
         hf_model_name="Trelis/Llama-2-7b-chat-hf-function-calling-v2",
         vmfb_path="Llama_2_7b_chat_hf_function_calling_v2.vmfb",
-        external_weight_file="Llama_2_7b_chat_hf_function_calling_v2_f16_int4.safetensors",
+        external_weight_file=f"Llama_2_7b_chat_hf_function_calling_v2_{precision}_{quantization}.safetensors",
         tokens_to_compare=50,
     )
+
+    torch_str = torch_str[: len(turbine_str)]
 
     import difflib
 
@@ -113,8 +114,3 @@ Be concise. You are a helpful, respectful and honest assistant. If a question do
     )
 
     assert torch_str == turbine_str, "".join(diff)
-
-
-if __name__ == "__main__":
-    logging.basicConfig(level=logging.DEBUG)
-    unittest.main()
